+++ lte/phy/include/liblte/phy/phch/dci.h
@@ -62,9 +62,9 @@ typedef enum {
 } dci_spec_t;
 
 typedef struct LIBLTE_API {
-  uint8_t nof_bits;
-  uint8_t L; // Aggregation level
-  uint8_t ncce; // Position of first CCE of the dci
+  uint32_t nof_bits;
+  uint32_t L; // Aggregation level
+  uint32_t ncce; // Position of first CCE of the dci
   uint16_t rnti;
 } dci_candidate_t;
 
@@ -75,20 +75,20 @@ typedef struct LIBLTE_API {
 
 typedef struct LIBLTE_API {
   dci_msg_t *msg;
-  uint8_t nof_dcis;
-  uint8_t max_dcis;
+  uint32_t nof_dcis;
+  uint32_t max_dcis;
 } dci_t;
 
 LIBLTE_API int dci_init(dci_t *q, 
-                        uint8_t max_dci);
+                        uint32_t max_dci);
 
 LIBLTE_API void dci_free(dci_t *q);
 
 LIBLTE_API char* dci_format_string(dci_format_t format);
 
 LIBLTE_API int dci_msg_candidate_set(dci_msg_t *msg, 
-                                     uint8_t L, 
-                                     uint8_t nCCE, 
+                                     uint32_t L, 
+                                     uint32_t nCCE, 
                                      uint16_t rnti);
 
 LIBLTE_API void dci_candidate_fprint(FILE *f, 
@@ -96,7 +96,7 @@ LIBLTE_API void dci_candidate_fprint(FILE *f,
 
 LIBLTE_API int dci_msg_get_type(dci_msg_t *msg, 
                                 dci_msg_type_t *type, 
-                                uint8_t nof_prb, 
+                                uint32_t nof_prb, 
                                 uint16_t crnti);
 
 LIBLTE_API void dci_msg_type_fprint(FILE *f, 
@@ -105,25 +105,25 @@ LIBLTE_API void dci_msg_type_fprint(FILE *f,
 // For dci_msg_type_t = PUSCH_SCHED
 LIBLTE_API int dci_msg_pack_pusch(ra_pusch_t *data, 
                                   dci_msg_t *msg, 
-                                  uint8_t nof_prb);
+                                  uint32_t nof_prb);
 
 LIBLTE_API int dci_msg_unpack_pusch(dci_msg_t *msg, 
                                     ra_pusch_t *data, 
-                                    uint8_t nof_prb);
+                                    uint32_t nof_prb);
 
 // For dci_msg_type_t = PDSCH_SCHED
 LIBLTE_API int dci_msg_pack_pdsch(ra_pdsch_t *data, 
                                   dci_msg_t *msg, 
                                   dci_format_t format, 
-                                  uint8_t nof_prb, 
+                                  uint32_t nof_prb, 
                                   bool crc_is_crnti);
 
 LIBLTE_API int dci_msg_unpack_pdsch(dci_msg_t *msg, 
                                     ra_pdsch_t *data, 
-                                    uint8_t nof_prb, 
+                                    uint32_t nof_prb, 
                                     bool crc_is_crnti);
 
-LIBLTE_API uint8_t dci_format_sizeof(dci_format_t format, 
-                                 uint8_t nof_prb);
+LIBLTE_API uint32_t dci_format_sizeof(dci_format_t format, 
+                                 uint32_t nof_prb);
 
 #endif // DCI_
